Move deep_check to common
